requirements.txt
python-dotenv
langchain-pinecone
cohere
langchain
langchain-community
langchain-core 
fastapi
uvicorn